Initial Marble software commit. Separate peripheral files. Each peripheral file includes code for its tasks. Lots of testing code present but commented. Will be removed after one successful run.
Not happy with TIM3 code, needs to be implemented in a better way or moved to adc peripheral. Moving to adc peripheral would be nice if a IIR filter is implemented on it. As it stands it is initial offset based.

Main control of the code is from TIM3 interrupt. Dosent interfere with UART sending.
Data being sent is just two values that are
Mag_ver_sensor
Mag_hor_sensor

Task remaining:
7 warnings
Clean up commented code!!!!!!
Check adc division!!!!!!!!
uart receiving needs to be added.
HAL_DELAY(1) Not needed.
Check execution of 0.1ms interrupt processing delay with GPIO

Code not tested on board!!